split non/kernel chainparams
